+++ standalone/scaler/scaler_client.cpp
@@ -108,7 +108,7 @@ uint64_t LoadRecordTime() {
 //-----------------------------------------------------------------------------
 
 void PrintVersion() {
-	printf("scaler_client 2.1.3\n");
+	printf("scaler_client 2.2.0\n");
 	printf("Part of the easy-config-logic, produced by pwl\n");
 	return;
 }
@@ -501,7 +501,7 @@ void HandleHistoryRequest(const httplib::Request &request, httplib::Response &re
 	// calculate start and end date
 	std::string start_date = body["start"];
 	std::string end_date = body["end"];
-	std::cout << start_date << " " << end_date << std::endl;
+	// std::cout << start_date << " " << end_date << std::endl;
 	tm start_time = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
 	strptime(start_date.c_str(), "%Y-%m-%d", &start_time);
 	time_t start_seconds = mktime(&start_time);
@@ -585,7 +585,6 @@ void HandleSettingsRequest(const httplib::Request &request, httplib::Response &r
 	response_json["status"] = 0;
 
 	if (body["request"] == "get-scaler-names") {
-		std::cout << "request get scaler names" << std::endl;
 		// read scaler names from config file
 		std::ifstream scaler_names_config("scaler-names.json");
 		if (!scaler_names_config.good()) {
@@ -598,19 +597,30 @@ void HandleSettingsRequest(const httplib::Request &request, httplib::Response &r
 			for (size_t i = 0; i < kScalerNum; ++i) {
 				response_json["names"].push_back(scaler_names["names"][i]);
 			}
-			std::cout << response_json.dump(2) << std::endl;
+			// std::cout << response_json.dump(2) << std::endl;
 		}
 		scaler_names_config.close();
 	} else if (body["request"] == "set-scaler-names") {
-		std::cerr << "Error: to do.\n";
-		response_json["status"] = 1;	
+		nlohmann::json scaler_names;
+		for (size_t i = 0; i < kScalerNum; ++i) {
+			scaler_names["names"].push_back(body["names"][i]);
+		}
+		std::ofstream scaler_names_config("scaler-names.json");
+		if (!scaler_names_config.good()) {
+			// error open file
+			std::cerr << "Error: open file scaler-names.json failed.\n";
+			response_json["status"] = 1;
+		}
+		scaler_names_config << scaler_names.dump(2) << std::endl;
+		scaler_names_config.close();
+		// std::cout << scaler_names.dump(2) << std::endl;
 	} else {
 		std::cerr << "Error: invalid request " << body["request"] << "\n";
 		response_json["status"] = 1;
 	}
 
 	response.set_content(response_json.dump(), "text/plain");
-	std::cout << "response " << response_json.dump(2) << std::endl;
+	// std::cout << "response " << response_json.dump(2) << std::endl;
 	return;
 }